feature: Implemented primitive player movement, boundaries and animation
Implemented the most basic player movement for two players, using WASD and Arrow Keys. The current movement system doesn't support gravity or any real physics system.
Added boundaries for said players, so that leaving the screen is currently impossible.
Added basic test sprites and a primitive way to animate movement and staying still.
The entire game logic is currently in the Main.cpp file, which should be refactored once an agreement on project structure is made.