feature: Improved player animations
Players now posess fully working running and idling animations, as well as the ability to change the direction they are looking at when they turn.
Frame speed of the animation has been tripled to be more noticable, which can in the future be changed in the Animations.cpp file.
Changing the player spawn order to be after all the other entities seemed to fix the lack of transparency issue, but the issue persists upon players picking up pickups.