feature: Overhauled combat system
Replaced the currently existing spellcasting system with a much more flexible one.
Added a bunch of enums that describe the properties of a certain spell, which can be added as components to spell entities. Based on these components, combat systems will do certain actions on these entities, like moving them or checking for certain collisions.
Separated the spell movement and collision systems, which now communicate with the use of signals.