refactor: Removed unnecessary code repeats and loop repeats with a slight change in game logic
Removed unnecessary Flip checks in various systems of Combat.cpp with a change of the spell range functionality and the addition of a new spell start position attribute.
Replaced order of loops for the pickup system to evade unnecessary picking up of player entities when no pick ups exist.
Changed name attributes of pickup and item components to evade potential memory leaks.