bugfix: Fixed bouncy players upon landing on a platform and changed how gravity and playercontrols systems work
Removed unnecessary jump_speed attribute of player component. Made it so that player vertical_velocity now works in both directions, to take account of both jumping and gravity.
Changed how platform collision works, so that it happens only if the player is falling (or standing still on a platform) and not while he is going up in a jump, which had to be done for the newly-implemented vertical_velocity to work.
All player position changes now happen in the PlayerControlsSystem, while GravitySystem only changes the vertical_velocity. This kind of kills the point of a gravity system existing separately, but with future plans of creating more entites that are affected by gravity, it should remain in existence.